refactor: collapse duplicated helpers left by the merged work quanta
Five quanta were developed in isolation, so a few concepts ended up
expressed more than once in the merged tree. None of these change
behaviour; each removes a copy a reader would otherwise have to hold.

WebAuthn production: `ZeroMemory(ReadOnlyMemory<byte>?)` existed
verbatim, comment and all, in both WebAuthnClient and WebAuthnBackend.
It is now `Util.SensitiveMemory.Zero`, next to the existing Base64Url
helper.

WebAuthn tests: MockFido2Responses is the shared fixture home, but
WebAuthnClientMakeCredentialTests still carried a byte-for-byte copy of
CreateMockMakeCredentialResponse and its three builders, and there were
three separate ways to build a mock GetAssertionResponse. The shared
builder now takes the optional numberOfCredentials / signature / user
that the private copies needed, and the copies are gone.

Core tests: GetNamedCurve and BuildUncompressedPoint each appeared in
two of the Asn*Tests files; both now live in EcTestSupport.
UncompressedPoint deliberately keeps its own implementation rather than
calling AsnUtilities.BuildUncompressedEcPoint, so encoder tests do not
check that helper against itself.
