+++ src/WebAuthn/src/Client/WebAuthnClient.Authentication.cs
@@ -21,6 +21,7 @@ using Yubico.YubiKit.WebAuthn.Client.Authentication;
 using Yubico.YubiKit.WebAuthn.Client.UserVerification;
 using Yubico.YubiKit.WebAuthn.Client.Validation;
 using Yubico.YubiKit.WebAuthn.Extensions;
+using Yubico.YubiKit.WebAuthn.Util;
 
 namespace Yubico.YubiKit.WebAuthn.Client;
 
@@ -201,7 +202,7 @@ public sealed partial class WebAuthnClient
         }
         finally
         {
-            ZeroMemory(request.PinUvAuthParam);
+            SensitiveMemory.Zero(request.PinUvAuthParam);
         }
     }
